requirements.txt
Flask>=1.0
Flask-WTF
Flask-Login
Flask-User
Flask-cache
Flask-script
Flask-moment
Flask-BabelEx
Flask-migrate
Flask-bootstrap
Flask-sqlalchemy
Flask-debugtoolbar
#uwsgi
redis
hiredis
fabric3
Pillow
gevent
selenium
validators
youtube_dl
SQLAlchemy
#sqlalchemy-utils
pytesseract
#MySQL-python, (not support py3)
mysqlclient
pyvirtualdisplay
selenium-requests
celery[gevent,redis]


